Add files via upload
dcgan.py is us trying to create our own dcgan based off of the medium article (https://medium.com/@liyin2015/dcgan-79af14a1c247)

Keras-DCGAN.py is a slightly modified program. Only difference is that it saves the plots and uses 'tk' as the event loop instead of 'inline'

modified_keras_gan.py is heavily modified from the given sample code. We did a lot of playing around with this one because it runs faster.